socialapi: code clean up for algolia connector
