+++ go/src/socialapi/workers/cmd/algoliaconnector/main.go
@@ -30,15 +30,16 @@ func main() {
 	// create message handler
 	handler := algoliaconnector.New(r.Log, algolia, appConfig.Algolia.IndexSuffix)
 	r.SetContext(handler)
-	r.Register(models.Channel{}).OnCreate().Handle((*algoliaconnector.Controller).TopicSaved)
-	r.Register(models.Channel{}).OnUpdate().Handle((*algoliaconnector.Controller).TopicUpdated)
-	r.Register(models.Account{}).OnCreate().Handle((*algoliaconnector.Controller).AccountSaved)
+	r.Register(models.Channel{}).OnCreate().Handle((*algoliaconnector.Controller).ChannelCreated)
+	r.Register(models.Channel{}).OnUpdate().Handle((*algoliaconnector.Controller).ChannelUpdated)
+	r.Register(models.Account{}).OnCreate().Handle((*algoliaconnector.Controller).AccountCreated)
+	r.Register(models.Account{}).OnUpdate().Handle((*algoliaconnector.Controller).AccountUpdated)
 	r.Register(models.ChannelMessageList{}).OnCreate().Handle((*algoliaconnector.Controller).MessageListSaved)
 	r.Register(models.ChannelMessageList{}).OnDelete().Handle((*algoliaconnector.Controller).MessageListDeleted)
 	r.Register(models.ChannelMessage{}).OnUpdate().Handle((*algoliaconnector.Controller).MessageUpdated)
 
 	// moderation related
-	r.Register(models.ChannelLink{}).OnCreate().Handle((*algoliaconnector.Controller).CreateSynonym)
+	r.Register(models.ChannelLink{}).OnCreate().Handle((*algoliaconnector.Controller).ChannelLinkCreated)
 
 	// participant related events
 	r.Register(models.ChannelParticipant{}).OnCreate().Handle((*algoliaconnector.Controller).ParticipantCreated)
